resolved null error on story component
