+++ src/App.js
@@ -1,4 +1,4 @@
-import { useEffect, useState } from 'react';
+// import { useEffect, useState } from 'react';
 import { Routes, Route } from 'react-router-dom';
 import './App.scss';
 import NavBar from './Components/NavBar.js';
@@ -8,37 +8,6 @@ import Story from './Components/Story';
 const Home = () => <h2>Home</h2>
 
 function App() {                    
-  const [stories, setStories] = useState('');
-  const storyData = stories;
-  // console.log(storyData);
-
-
-  useEffect(() => {
-
-    const fetchData = async () => {
-      const apiKey = 'n92DZWsbGi92AGqOU61WrbTbzwcLBSKT';
-      const url = new URL(`https://api.nytimes.com/svc/topstories/v2/home.json?api-key=${apiKey}`)
-
-      try {
-        const data = await fetch(url);
-        const response = await data.json();
-        const resultsArray = response;
-        // console.log(response);
-        setStories(resultsArray);
-        // console.log(stories);
-
-      } catch (error) {
-        console.log(error);
-      }
-    } 
-    
-    fetchData();
-    
-
-  },[]);
-  // console.log(stories);
-  
-  
 
   return (
     <div className="App">
@@ -54,7 +23,7 @@ function App() {
         </div>
       </header>
             <h2>Breaking News</h2>
-            <Story data={ storyData } />
+            <Story />
