+++ src/partials/_story.scss
@@ -2,6 +2,7 @@
 .storyComponent {
     display: flex;
     justify-content: center;
+    flex-direction: column;
 }
 .storyBox {
     display: flex;
